requirements.txt
watchdog>=3.0.0
pystray>=0.19.0
Pillow>=9.0.0
pyinstaller>=5.0.0
